Complete attach runtime and interactive layout integration
- refactor prompt attach orchestration around measured layouts, scoped
presentation resources, and typed input routing
- make search/select scrolling row-aware, resize-stable, and
scrollbar-aware
- improve text editing, modal/picker composition, flex layout, pane
clipping, and component geometry
- preserve window discovery and tab-strip context routing across narrow
responsive layouts
- expand playbook coverage for real attach navigation and context
switching

+++ packages/tui-components/src/button.rs
@@ -257,7 +257,7 @@ impl Component for ButtonComponent<'_, '_> {
             .visible_rect(bmux_tui::component::LogicalRect::new(
                 0,
                 0,
-                layout.size.width,
+                usize::from(layout.size.width),
                 1,
             ))
             .intersection(area);
